api/search_youtube/transcription.py
import os

from deepgram import DeepgramClient, FileSource, PrerecordedOptions

API_KEY = os.getenv("DEEPGRAM_SECRET_KEY")


def convert_audio_to_text(audio_file):
    try:
        deepgram = DeepgramClient(API_KEY)

        with open(audio_file, "rb") as file:
            buffer_data = file.read()

        payload: FileSource = {
            "buffer": buffer_data,
        }

        options = PrerecordedOptions(
            model="nova-2",
            smart_format=True,
        )

        response = deepgram.listen.prerecorded.v("1").transcribe_file(payload, options)

        return response["result"]["channels"][0]["alternatives"][0]["transcript"]

    except Exception as e:
        print(f"Exception: {e}")
